refactor: fold quickpay coordinator into repo
